Improve pricer and move coingecko to separate mod
