--- .gitattributes
@@ -1,1 +1,0 @@
-recipe.json filter=lfs diff=lfs merge=lfs -text
