+++ apps/pricer/src/coingecko/mapper.rs
@@ -26,7 +26,7 @@ pub fn get_chain_for_coingecko_id(id: &str) -> Option<Chain> {
         "celestia" => Some(Chain::Celestia),
         "injective-protocol" => Some(Chain::Injective),
         "sei-network" => Some(Chain::Sei),
-        "manta-network" => Some(Chain::Manta),
+        "manta-network" | "manta-pacific" => Some(Chain::Manta),
         _ => None,
     }
 }
