+++ apps/pricer/src/coingecko/mod.rs
@@ -1,0 +1,5 @@
+pub mod client;
+pub mod mapper;
+pub mod model;
+pub use self::client::CoinGeckoClient;
+pub use self::model::{CoinInfo, CoinMarket};
