+++ apps/pricer/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod client;
 pub mod coingecko;
-pub mod price_mapper;
 pub mod price_updater;
 
 pub const DEFAULT_FIAT_CURRENCY: &str = "USD";
