--- recipe.json
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:e125a17a5512eaa222c35caf795774f370236188384f5c58f083ca0040929ae6
-size 228736
